Pull request comments: Remove verbose documentation. Renamed commands according to convention (CmdXxxx).
